Перенёс функции ассоциации в ассоциатор
