+++ vk_maria/keyboard/associator.py
@@ -1,23 +1,53 @@
-from .utils import get_keyboards_from_files, get_keyboards_from_models
+import inspect
+import json
+import os
+import sys
+
 from .exceptions import *
+from .model import construct_json, Model
+from .. import working_dir
 
 
 class KeyboardAssociator:
     """
     Объект, через который можно обращаться к клавиатурам из моделей или папки.
 
-    :param models: Файл с моделями клавиатур.
+    :param module: Файл с моделями клавиатур.
     :param folder: Папка с файлами json клавиатур.
     """
 
-    def __init__(self, models: str = None, folder: str = None):
-        self.KEYBOARDS_MODELS = models
+    @staticmethod
+    def get_keyboards_from_files(folder):
+        keyboard_dir = os.path.join(working_dir, folder)
+
+        keyboards = {}
+        for file in os.listdir(keyboard_dir):
+            keyboards[file.split('.')[0]] = open(os.path.join(keyboard_dir, file)).read()
+
+        return keyboards
+
+    @staticmethod
+    def get_keyboards_from_models(models):
+        module = __import__(models)
+        keyboard_models = []
+        for keyboard in inspect.getmembers(sys.modules[module.__name__], inspect.isclass):
+            if Model in keyboard[1].__bases__:
+                keyboard_models.append(keyboard[1])
+
+        keyboards = {}
+        for keyboard in keyboard_models:
+            keyboards[keyboard.__name__] = json.dumps(construct_json(keyboard))
+
+        return keyboards
+
+    def __init__(self, module: str = None, folder: str = None):
+        self.KEYBOARDS_MODELS = module
         self.KEYBOARDS_FOLDER = folder
         self.keyboards = {}
         if folder:
-            self.keyboards.update(get_keyboards_from_files(folder))
-        if models:
-            self.keyboards.update(get_keyboards_from_models(models))
+            self.keyboards.update(self.get_keyboards_from_files(folder))
+        if module:
+            self.keyboards.update(self.get_keyboards_from_models(module))
 
     def __getitem__(self, keyboard_name):
         try:
